functions/voice/pyproject.toml
[tool.poetry]
name = "voice"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false
packages = [{include = "voice"}]

[tool.poetry.dependencies]
python = "^3.11"
firebase-admin = "7.1.0"
google-cloud-texttospeech = "2.29.0"
google-cloud-speech = "2.33.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
